requirements.txt
aiofiles==23.2.1
requests
altair==5.3.0
annotated-types==0.6.0
anyio==4.4.0
asttokens==2.4.1
attrs==23.2.0
Automat==22.10.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
Bottleneck==1.4.0
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
constantly==23.10.4
contourpy==1.2.1
cryptography==42.0.7
cssselect==1.2.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
executing==2.1.0
fake-useragent==1.5.1
fastapi==0.111.0
fastapi-cli==0.0.3
ffmpy==0.3.2
filelock==3.14.0
fire==0.6.0
fonttools==4.51.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
gradio==4.40.0
gradio_client==1.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.7
importlib_resources==6.4.0
incremental==22.10.0
ipykernel==6.29.5
ipython==8.27.0
itemadapter==0.9.0
itemloaders==1.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
langfuse==2.39.3
linkedin-scraper==2.11.2
loguru==0.7.2
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numexpr==2.10.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openai==1.28.1
opencv-python==4.9.0.80
openvino==2024.1.0
openvino-telemetry==2024.1.0
orjson==3.10.3
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.2
parsel==1.9.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.3.2
playwright==1.46.0
prompt_toolkit==3.0.47
Protego==0.3.1
proto-plus==1.24.0
protobuf==5.27.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybboxes==0.1.6
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
PyDispatcher==2.0.7
pydub==0.25.1
pyee==11.1.0
Pygments==2.18.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.2.0
queuelib==1.7.0
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
ruff==0.4.4
sahi==0.11.18
scikit-learn==1.5.2
scipy==1.14.1
Scrapy==2.11.1
scrapy-playwright==0.0.41
scrapy-splash==0.9.0
seaborn==0.13.2
selenium==4.22.0
semantic-version==2.10.0
service-identity==24.1.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
slack_sdk==3.31.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
starlette==0.37.2
streamlit==1.38.0
sympy==1.13.2
tenacity==8.5.0
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tiktoken==0.7.0
tldextract==5.1.2
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
trio==0.26.0
trio-websocket==0.11.1
triton==3.0.0
Twisted==24.3.0
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
ultralytics==8.2.92
ultralytics-thop==2.0.6
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
validators==0.33.0
w3lib==2.1.2
watchdog==4.0.2
watchfiles==0.21.0
wcwidth==0.2.13
webdriver-manager==4.0.1
websocket-client==1.8.0
websockets==11.0.3
wrapt==1.16.0
wsproto==1.2.0
zope.interface==6.3
